refactoring EdgeAttr definitions to avoid cumbersome typeindex comparisons; streamlining packing/unpacking of attributes in scatter_graph; bug fix in transmission of numbers of attributes in scatter_graph

+++ include/io/hdf5/attributes.hh
@@ -66,7 +66,7 @@ namespace ngh5
       herr_t num_edge_attributes
       (
        const std::vector< std::pair<std::string,hid_t> >& attributes,
-       std:: vector <size_t> &num_attrs
+       std:: vector <uint32_t> &num_attrs
        );
 
       /// @brief Reads the values of edge attributes.
